+++ hmily-tac/hmily-tac-sqlparser/hmily-tac-sqlparser-model/src/main/java/org/dromara/hmily/tac/sqlparser/model/segment/dml/expr/simple/HmilyParameterMarkerExpressionSegment.java
@@ -27,7 +27,7 @@ import lombok.ToString;
 @RequiredArgsConstructor
 @Getter
 @ToString
-public class HmilyParameterMarkerHmilyExpressionSegmentHmily implements HmilySimpleHmilyExpressionSegment {
+public class HmilyParameterMarkerExpressionSegment implements HmilySimpleExpressionSegment {
     
     private final int startIndex;
